now lets run our application
We'll mark our GreetingResource as authenticated now.
We'll now set the oidc provider to `google` because we want to use
auth from GCloud. We also need to obtain the client_id from GCloud.

Quarkus provides an env variable to set the client ID as well, but
that is long in name, so we can use the `CLIENTID` env variable instead
for convenience.

CLIENTID=xxx.apps.googleusercontent.com PORT=9000 ./gradlew quarkusDev

Now when you open the application then your google auth will show up
and give a redirect URI error.